+++ src/__tests__/Atom.test.ts
@@ -1,5 +1,5 @@
-import Atom from '../Atom';
-import IndexMap from '../IndexMap';
+import Atom from '../core/Atom';
+import IndexMap from '../core/IndexMap';
 import getAtom from '../utils/factories/Atom';
 import getAtomId from '../utils/factories/AtomId';
 import getAtomValue from '../utils/factories/AtomValue';
