+++ src/core/IndexMap.ts
