+++ src/ui/Cursor.d.ts
